discord-bot/requirements.txt
discord.py>=1.7.3,<1.8
pyyaml>=5.3.1,<5.4
